chore: Code structure & formatting for `@/components/navigation`
- Merged contents of `Back.tsx` & `BackButton.tsx`
