+++ src/app/_layout.tsx
@@ -6,8 +6,8 @@ import { useLoadAssets } from "@/hooks/useLoadAssets";
 import "@/assets/global.css";
 import { PlaybackService } from "@/constants/PlaybackService";
 import { AppProvider } from "@/providers/app";
-import { Header } from "@/components/navigation/Header";
-import { AnimatedBootSplash } from "@/components/navigation/AnimatedBootSplash";
+import { Header } from "@/components/navigation/header";
+import { AnimatedBootSplash } from "@/components/navigation/animated-boot-splash";
 import { AppModals } from "@/features/modal";
 
 export {
