--- src/components/navigation/Back.tsx
@@ -1,11 +1,0 @@
-import { router } from "expo-router";
-import { useEffect } from "react";
-
-/** @description Navigate back when rendered. */
-export function Back() {
-  useEffect(() => {
-    router.back();
-  }, []);
-
-  return null;
-}
