+++ src/components/navigation/animated-boot-splash.tsx
