+++ src/components/navigation/external-link.tsx
